# Conflicts:
#	grid2op/Runner/Runner.py

+++ grid2op/tests/test_notebooks_getting_started.py
@@ -1,19 +1,20 @@
-import os
-import shutil
-import sys
-import unittest
-import json
-import numpy as np
-import subprocess
+# Copyright (c) 2019-2020, RTE (https://www.rte-france.com)
+# See AUTHORS.txt
+# This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
+# If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
+# you can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
+# This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
+import shutil
 import pdb
 
-import nbformat
-from nbconvert.preprocessors import ExecutePreprocessor, CellExecutionError
-
-from helper_path_test import PATH_DATA_TEST
+from grid2op.tests.helper_path_test import *
+import os
+import unittest
 
 # TODO check these tests, they don't appear to be working
+
 
 def delete_all(folder):
     """
@@ -69,30 +70,34 @@ class TestNotebook(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    notebooks_path = os.path.abspath(os.path.join(PATH_DATA_TEST, "../../getting_started"))
-    path_save_notebook = os.path.join(PATH_DATA_TEST, "../tests/test_notebooks")
+    if False:
+        import nbformat
+        from nbconvert.preprocessors import ExecutePreprocessor, CellExecutionError
+        notebooks_path = os.path.abspath(os.path.join(PATH_DATA_TEST, "../../getting_started"))
+        path_save_notebook = os.path.join(PATH_DATA_TEST, "../tests/test_notebooks")
 
-    if not os.path.exists(path_save_notebook):
-        os.mkdir(path_save_notebook)
-    else:
-        delete_all(path_save_notebook)
+        if not os.path.exists(path_save_notebook):
+            os.mkdir(path_save_notebook)
+        else:
+            delete_all(path_save_notebook)
 
-    all_notebook = export_all_notebook(notebooks_path)
+        all_notebook = export_all_notebook(notebooks_path)
 
-    all_funs = []
-    for notebook_filename in all_notebook:
-        def f(self):
-            with open(notebook_filename) as f:
-                nb = nbformat.read(f, as_version=4)
-                ep = ExecutePreprocessor(timeout=600)
-                try:
-                    ep.preprocess(nb, {'metadata': {'path': path_save_notebook}})
-                except CellExecutionError:
-                    raise
-        all_funs.append(("test_{}".format(os.path.split(notebook_filename)[-1]), f))
+        all_funs = []
+        for notebook_filename in all_notebook:
+            def f(self):
+                with open(notebook_filename) as f:
+                    nb = nbformat.read(f, as_version=4)
+                    ep = ExecutePreprocessor(timeout=600)
+                    try:
+                        ep.preprocess(nb, {'metadata': {'path': path_save_notebook}})
+                    except CellExecutionError:
+                        raise
+            all_funs.append(("test_{}".format(os.path.split(notebook_filename)[-1]), f))
 
-    for nm, f in all_funs:
-        setattr(TestNotebook, nm, f)
+        for nm, f in all_funs:
+            pass
+            setattr(TestNotebook, nm, f)
 
-    unittest.main()
+    # unittest.main()
